third_party/blink/web_tests/wpt_internal/scheduler/task_result_error.html

<!doctype html>
<title>Scheduling API: Task.result When a Task Throws an Erorr</title>
<link rel="author" title="Scott Haseley" href="mailto:shaseley@chromium.org">
<link rel="help" href="https://github.com/WICG/main-thread-scheduling">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
'use strict';

// Uncaught exceptions in asynchronous tasks are still reported to the global
// handler, which we expect to happen in this test.
setup({allow_uncaught_exception : true});

async_test(t => {
  (function() {
    let task = scheduler.postTask(() => { throw Error('Failed'); });
    task.result.then(t.step_func((res) => {
      assert_true(false, 'task.result should not be fulfilled when the task throws an error.');
    }))
    .catch(t.step_func_done((e) => {
      assert_equals(e.name, 'Error');
      assert_equals(e.message, 'Failed');
    }));
  })();
}, 'Test task.result is rejected properly when the task throws an error');

</script>
